+++ robotiq_2_finger_gripper_driver/CMakeLists.txt.ros2
@@ -50,11 +50,15 @@ ament_target_dependencies(${PROJECT_NAME}_components rclcpp rclcpp_components)
 rosidl_get_typesupport_target(${PROJECT_NAME}_cpp_msgs_target ${PROJECT_NAME}_msgs rosidl_typesupport_cpp)
 target_link_libraries(${PROJECT_NAME}_components ${PROJECT_NAME} ${${PROJECT_NAME}_cpp_msgs_target})
 
-rclcpp_components_register_node(
+## Register C++ components
+rclcpp_components_register_nodes(
   ${PROJECT_NAME}_components
-  PLUGIN "robotiq_2_finger_gripper_driver::Robotiq2FingerDriverNode"
-  EXECUTABLE ${PROJECT_NAME}_node
+  "robotiq_2_finger_gripper_driver::Robotiq2FingerDriverNode"
 )
+
+## Declare C++ component executable
+add_executable(${PROJECT_NAME}_node src/robotiq_2_finger_gripper_driver_node_main.ros2.cpp)
+target_link_libraries(${PROJECT_NAME}_node ${PROJECT_NAME}_components)
 
 #############
 ## Install ##
@@ -65,6 +69,11 @@ install(TARGETS ${PROJECT_NAME} ${PROJECT_NAME}_components
   ARCHIVE DESTINATION lib
   LIBRARY DESTINATION lib
   RUNTIME DESTINATION bin
+)
+
+install(
+  TARGETS ${PROJECT_NAME}_node
+  RUNTIME DESTINATION lib/${PROJECT_NAME}
 )
 
 ## Mark cpp header files for installation
